menu is working, weekend section is scraping and adding/deleting from favs

+++ lib/activity.rb
@@ -5,27 +5,34 @@ require 'tty-prompt'
 require 'colorize'
 require 'json'
 
+#the Class Activity is the application
 class Activity
 
-    attr_reader :today_activities, :fav_list, :chosen_today_activity, :processed_today
-    attr_accessor :selected_activity
+    attr_reader :today_activities, :fav_list, :chosen_today_activity, :processed_today, :weekend_activities
+    attr_accessor :selected_activity, :active
 
 # Initialise class variables and scrape website for current information
     def initialize(file_path)
         @today_activities = []
+        @weekend_activities = []
         @processed_today = []  
+        @processed_weekend = []  
         @processed_favs = []
         scrape_today
+        scrape_weekend
         @chosen_today_activity
+        @chosen_weekend_activity
         @selected_activity =[]
         @fav_list = []
         @prompt = TTY::Prompt.new
-        @header = "-"*50
+        @header = "-"*60
         @header_length = @header.length
         process_today_activities
+        process_weekend_activities
         @check = true
         @file_path = file_path
         load_data(file_path)
+       
         
            
     end
@@ -53,6 +60,20 @@ class Activity
             @today_activities << activity                
         end               
     end
+# Method to scrape website for whats happening on the weekend in Melbourne
+    def scrape_weekend
+        url = "https://www.timeout.com/melbourne/things-to-do/things-to-do-in-melbourne-this-weekend"
+        unparsed_page = HTTParty.get(url)
+        parsed_page = Nokogiri::HTML(unparsed_page.body)
+        items = parsed_page.css('div.card-content')
+        items.each do |activity|
+            activity = {
+                title: activity.css('h3.card-title').text.strip,
+                description: activity.css('p').text
+            }  
+            @weekend_activities << activity                
+        end               
+    end
 
 # Method to display today titles
     def display_today_activity_name
@@ -76,10 +97,11 @@ class Activity
 
 #Method to display the main menu
     def main_menu
+        @active = 0
         @prompt.select("What would you like to do?") do |menu|
             menu.choice "Check out what's on today", -> {today_menu} 
-            menu.choice "Check out things to do on the weekend", -> {}
-            menu.choice "Look at my favorites", -> {display_favorites}
+            menu.choice "Check out things to do on the weekend", -> {weekend_menu}
+            menu.choice "Look at my favourites", -> {display_favorites}
             menu.choice "Exit", -> {leave_app}    
             end    
     end
@@ -99,14 +121,19 @@ class Activity
                     @selected_activity << today_activities[chosen_today_activity-1][:title] 
                     
                     today_return_menu
+        else
+            if chosen_today_activity == 11
+                welcome_header
+                main_menu
+            end
         end
         
     end
 
 #Method to display today return menu
     def today_return_menu
-        today_return_selection = @prompt.select("\nSave to your favorites list or go back".blue) do |menu|
-            menu.choice "Save to favorites", 1
+        today_return_selection = @prompt.select("\nSave to your favourites list or go back".blue) do |menu|
+            menu.choice "Save to favourites", 1
             menu.choice "Go back", 2
            end 
            today_return_selection(today_return_selection)  
@@ -122,26 +149,33 @@ class Activity
                 favorite_saved 
                 fav_saved_return   
             else
-                puts "\nYou already have #{@selected_activity.last} in your favorites".red
+                puts "\nYou already have #{@selected_activity.last} in your favourites".red
                 fav_saved_return   
             end          
         when 2
+            if active == 1 
             today_menu
+            else   
+            weekend_menu
+            end
+           
         else
-            puts "Thats not right..."
+            puts "That's not right..."
         end   
     end
 
 #Method to verify favorite is saved
     def favorite_saved
-        puts "\n#{@selected_activity.last} saved to favorites!".green      
+        puts "\n#{@selected_activity.last} saved to favourites!".green      
     end
 
 #Method to display options after favorite saved
     def fav_saved_return
         fav_return_selection = @prompt.select("\nWhat would you like to do".blue) do |menu|
-            menu.choice "Return to todays activities", 1
-            menu.choice "Go to main menu", 2
+            menu.choice "See today activities", 1
+            menu.choice "See weekend activities", 2
+            menu.choice "See my favourites", 3
+            menu.choice "--Main Menu--", 4
            end
            fav_saved_selection(fav_return_selection)      
     end
@@ -152,10 +186,12 @@ class Activity
         when 1
             today_menu           
         when 2
+            weekend_menu
+        when 3
+            display_favorites
+        else
             welcome_header
             main_menu
-        else
-            puts "Thats not right..."
         end   
         
     end
@@ -166,7 +202,7 @@ class Activity
     def display_favorites
         favorites_header
         if @fav_list.length == 0
-            puts "You have no favorites!!!".yellow
+            puts "You have no favourites!!!".yellow
         else
         puts @fav_list.uniq
         end
@@ -178,7 +214,7 @@ class Activity
     def favorites_menu
         
         process_favorites
-        deleted_favorite = @prompt.select("\nSelect a favorite to delete or EXIT to main menu\n", @processed_favs.uniq.push({name: "--EXIT--", value: fav_list.length+1})) 
+        deleted_favorite = @prompt.select("\nSelect a favourite to delete or EXIT to main menu\n", @processed_favs.uniq.push({name: "--EXIT--".green, value: fav_list.length+1}), active_color: :red) 
        delete_selection(deleted_favorite)
     end
 
@@ -186,9 +222,18 @@ class Activity
     def delete_selection(deleted_favorite)
 
         if deleted_favorite <= @fav_list.length
-            delete_favorite(deleted_favorite)
+                delete = @prompt.ask("\nAre you sure?\nEnter 'Y' to confirm or any key to return") do |q|
+                q.modify :down
+              end
+              case delete  
+              when "y"
+                delete_favorite(deleted_favorite)
             @processed_favs =[] 
             display_favorites
+            else
+                display_favorites
+            end
+            
         else 
         welcome_header
         main_menu
@@ -197,15 +242,26 @@ class Activity
     end
    
 
-
-
-#Method to process the activity list for the menu
+#Method to process the today activity list for the menu
     def process_today_activities
         i=0
         while i < 10
                item  = {name: today_activities[i][:title], value: i+1.to_i}  
                 
                 @processed_today << item
+                 
+                i += 1     
+            end   
+            
+    end
+
+#Method to process the weekend activity list for the menu
+    def process_weekend_activities
+        i=0
+        while i < 10
+               item  = {name: weekend_activities[i][:title], value: i+1.to_i}  
+                
+                @processed_weekend << item
                  
                 i += 1     
             end   
@@ -222,50 +278,86 @@ class Activity
                  
                 i += 1     
             end   
-            
     end
 
+#Method to display the weekend menu
+    def weekend_menu
+        active = true
+        weekend_header
+        @selected_activity = []
+        chosen_weekend_activity = @prompt.select("\nSelect an activity to find out more\n", @processed_weekend.push({name: "--Back to Menu--", value: 11}).uniq) 
+        weekend_selection(chosen_weekend_activity)    
+    end
 
+ #Method to receive the weekend selection
+    def weekend_selection(chosen_weekend_activity)
+        if  chosen_weekend_activity <= 10
+                    puts "\n#{weekend_activities[chosen_weekend_activity-1][:description]}"
+                    @selected_activity << weekend_activities[chosen_weekend_activity-1][:title] 
+                    
+                    today_return_menu
+        else
+            if chosen_weekend_activity == 11
+                welcome_header
+                main_menu
+            end
+        end
+        
+    end
+    
+ #Method to display the welcome header
     def welcome_header
         system 'clear'
         puts @header.light_blue
         puts "Lets find something to do!".center(@header_length)
         puts @header.light_blue
     end
-
+#Method to display the today section header
     def today_header
         system 'clear'
         puts @header.light_blue
-        puts "Find something to do today!".center(@header_length)
+        puts "Check out whats on around Melbourne today!".center(@header_length)
         puts @header.light_blue
     end
-
+#Mthod to display the weekend header
+    def weekend_header
+        system 'clear'
+        puts @header.light_blue
+        puts "Check out whats on around Melbourne this weekend!".center(@header_length)
+        puts @header.light_blue
+    end
+#Method to display the favourites header
     def favorites_header
         system 'clear'
         puts @header.light_blue
-        puts "My Favorites".center(@header_length)
+        puts "My Favourites".center(@header_length)
         puts @header.light_blue
     end
-
+#method to display the leave header
     def leave_header
         system 'clear'
         puts @header.light_blue
         puts "See you next time!".center(@header_length)
         puts @header.light_blue
     end
-
+#Method to display the line divider
     def line_divider   
         puts @header.light_blue
     end
+#Method to clear the screen
+    def clear
+        system 'clear'
+        
+    end
 
-
-#Method to check if favorite exists
+#Method to check if favorite exists in the list
 
     def favorite_check
         @check = @fav_list.include?(@selected_activity.last)
        
     end
 
+#Method to leave the app
     def leave_app
         leave_header
         puts "Have a great day!".center(@header_length).blue
@@ -275,6 +367,7 @@ class Activity
         exit!   
     end
 
+#Method to load saved data
     def load_data(file_path)
         json_data = JSON.parse(File.read(file_path))
         @processed_favs = json_data.map do |fav|
@@ -291,76 +384,5 @@ class Activity
         end
     end
 
- 
-
 end
-
-# Title blocksheaders
-
-
-
-
-
-    # def lists_today_activities
-    #     i =0
-    #      while i < today_activities.length-1
-    #         title = today_activities[i][:title]
-    #         puts "#{title.strip}"
-            
-    #         i += 1
-    #      end
-    #     end
-
-    #     def combine_today_titles
-    #         i =0
-    #          while i < today_activities.length-1
-    #             title = today_activities[i][:title]
-                
-    #             @today_titles << title
-    #             i += 1
-    #          end
-    #     end
-
-    #     def combine_today_description
-    #         i =0
-    #          while i < today_activities.length-1
-    #             description = today_activities[i][:description]
-                
-    #             @combine_today_description << description
-    #             i += 1
-    #         end
-            
-    #     end
-
-    #     def activity_display
-    #         i=0
-    #        while i < 10
-               
-    #           bob = {name: today_activities[i][:title], value: i+1.to_i}
-
-    #           @list << bob
-
-    #             i += 1     
-    #         end
-    #     end 
-
-    #     def melb_list_today
-
-    #         i=0
-    #        while i < 10
-               
-    #           bob = {name: today_activities[i][:title], value: i+1.to_i}           
-    #              @list << bob
-
-    #             i += 1     
-    #         end
-        
-    #     end        
-    
-
-
-
-
-
-
